backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pandas==2.1.4
polars==0.20.0
celery==5.3.4
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
openpyxl==3.1.2
aiosqlite==0.19.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
python-dateutil==2.8.2